+++ src/main/java/com/concord/petmily/auth/service/TokenService.java
@@ -3,7 +3,7 @@ package com.concord.petmily.auth.service;
 import com.concord.petmily.auth.exception.AuthException;
 import com.concord.petmily.common.exception.ErrorCode;
 import com.concord.petmily.user.entity.User;
-import com.concord.petmily.user.service.UserService;
+import com.concord.petmily.user.service.UserServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -15,7 +15,7 @@ public class TokenService {
 
     private final TokenProvider tokenProvider;
     private final RefreshTokenService refreshTokenService;
-    private final UserService userService;
+    private final UserServiceImpl userServiceImpl;
 
     public String createNewAccessToken(String refreshToken) {
         if (!tokenProvider.validToken(refreshToken)) {
@@ -23,7 +23,7 @@ public class TokenService {
         }
 
         String username = refreshTokenService.findByRefreshToken(refreshToken).getUsername();
-        User user = userService.findByUsername(username);
+        User user = userServiceImpl.findByUsername(username);
 
         return tokenProvider.generateToken(user, Duration.ofHours(2));
     }
